requirements.txt
torch~=1.3
torchvision~=0.4
pytorch-ignite~=0.2
tensorboard~=2.0
numpy~=1.17
pandas~=0.25
scikit-learn~=0.21
scikit-video~=1.1
tqdm~=4.37
jupyterlab~=1.2
seaborn~=0.9
bokeh~=1.4
ffmpeg~=1.4